Move E2E tests for Swaps into its own folder, move shared things into its own shared.js file
