+++ test/e2e/swaps/swap-eth.spec.js
@@ -1,58 +1,23 @@
 const { strict: assert } = require('assert');
+
 const { withFixtures } = require('../helpers');
 const FixtureBuilder = require('../fixture-builder');
+const { ganacheOptions, loadSwaps, buildQuote } = require('./shared');
 
 describe('Swap Eth for another Token', function () {
-  const ganacheOptions = {
-    accounts: [
-      {
-        secretKey:
-          '0x7C9529A67102755B7E6102D6D950AC5D5863C98713805CEC576B945B15B71EAC',
-        balance: 25000000000000000000,
-      },
-    ],
-  };
   it('Completes a Swap between Eth and Matic', async function () {
     await withFixtures(
       {
         fixtures: new FixtureBuilder().build(),
         ganacheOptions,
         title: this.test.title,
-        failOnConsoleError: false,
-        driverOptions: {
-          timeOut: 20000,
-        },
       },
       async ({ driver }) => {
-        await driver.navigate();
-        await driver.fill('#password', 'correct horse battery staple');
-        await driver.press('#password', driver.Key.ENTER);
-        await driver.clickElement(
-          '.wallet-overview__buttons .icon-button:nth-child(3)',
-        );
-        await driver.clickElement(
-          '[class*="dropdown-search-list"] + div[class*="MuiFormControl-root MuiTextField-root"]',
-        );
-        await driver.fill('input[placeholder*="0"]', '2');
-        await driver.clickElement(
-          '[class*="dropdown-search-list"] + div[class*="MuiFormControl-root MuiTextField-root"]',
-        );
-        await driver.clickElement(
-          '[class="dropdown-search-list__closed-primary-label dropdown-search-list__select-default"]',
-        );
-        await driver.clickElement(
-          '[placeholder="Search name or paste address"]',
-        );
-        await driver.fill(
-          '[placeholder="Search name or paste address"]',
-          'DAI',
-        );
-        await driver.waitForSelector(
-          '[class="searchable-item-list__primary-label"]',
-        );
-        await driver.clickElement(
-          '[class="searchable-item-list__primary-label"]',
-        );
+        await loadSwaps(driver);
+        await buildQuote(driver, {
+          amount: 2,
+          swapTo: 'DAI',
+        });
         await driver.clickElement({ text: 'Review swap', tag: 'button' });
         await driver.waitForSelector('[class*="box--align-items-center"]');
         const estimatedEth = await driver.waitForSelector({
